+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* [Issue #194](https://github.com/manheim/terraform-pipeline/issues/194) Add support global tfvars files.
+
 # v5.4
 
 * [Issue #87](https://github.com/manheim/terraform-pipeline/issues/87) Add travis CI support to the project